Revert "Code cleanup"
This reverts commit 69643f704cadf157cd865389707b9e187e31aac6.

+++ src/main/java/network/crypta/clients/http/WelcomeToadlet.java
@@ -803,7 +803,7 @@ public class WelcomeToadlet extends Toadlet {
                 new String[] {"fullVersion", "build", "rev"},
                 new String[] {
                   Long.toString(Version.currentBuildNumber()),
-                  Integer.toString(Version.currentBuildNumber()),
+                  Integer.toString((int) Version.currentBuildNumber()),
                   Version.gitRevision()
                 }));
     versionContent.addChild("br");
