+++ src/main/java/network/crypta/node/updater/NodeUpdater.java
@@ -128,7 +128,7 @@ public abstract class NodeUpdater implements ClientGetCallback, USKCallback, Req
         try {
           manager
               .getUpdateOverMandatory()
-              .processMainJarBlob(temp, null, currentVersion, uri);
+              .processMainJarBlob(temp, null, (long) currentVersion, uri);
         } catch (Throwable t) {
           // Don't disrupt startup.
           Logger.error(this, "Unable to process old blob, caught " + t, t);
